+++ DU-Industry-Tool/SchematicValueForm.Designer.cs
@@ -77,7 +77,7 @@
             this.button1.TabIndex = 1;
             this.button1.Text = "&Save";
             this.button1.UseVisualStyleBackColor = true;
-            this.button1.Click += new System.EventHandler(this.button1_Click);
+            this.button1.Click += new System.EventHandler(this.Button1_Click);
             // 
             // SchematicName
             //