+++ DU-Industry-Tool/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.7.3
+
+- Added display of Industry below Products with tier and size.
+- Added "Industry" attribute to all 3203 recipes.
+- Removed all "optical sensor" and several "lights" parts as these are not available in game.
+- Code reviewed to have no more warnings in VS2019 IntelliSense list.
+
 ## v0.7.2
 
 - Improved search in tree to prioritize exact match over partial match.